% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources_fn_biotic.R
\name{abbreviate_taxa}
\alias{abbreviate_taxa}
\title{Abreviar Nombres Taxonómicos con Desambiguación}
\usage{
abbreviate_taxa(taxa_name)
}
\arguments{
\item{taxa_name}{Vector de caracteres con los nombres científicos.}
}
\value{
Un vector de caracteres con las abreviaciones.
}
\description{
Genera códigos abreviados. Si detecta colisiones (mismo código para especies distintas), aumenta la longitud del género para diferenciarlas.
}
