% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources_fn_abiotic.R
\name{fn_glm}
\alias{fn_glm}
\title{fn_glm}
\usage{
fn_glm(data, col_pars, col_valor, col_factor)
}
\arguments{
\item{data}{archivo entrada que tiene que tener, al menos, las columnas: sitio, parámetros (nombre o sigla), valor de parámetros. Tiene que estar en formato "long". Se recomida usar csv o tsv como formatos.}

\item{col_pars}{string que indica el nombre de la columna en data que tiene los parámetros}

\item{col_valor}{string que indica el nombre de la columna que tiene el valor de los parámetros.}

\item{col_factor}{string que indica el factor para ajustar modelo glm}
}
\value{
tablas y gráfios resultantes del modelo glm
}
\description{
función para ajustar modelo glm para evaluar efecto de un factor(es) sobre la concenytraciób de parámetros fisicoquiímicos.
}
\examples{
\dontrun{
# figura sin grupos
data <- readr::read_tsv("data.tsv")
col_pars <- "Param"
col_valor <- "Resultado"
col_factor <- "Campaña"
fn_glm(data = data, col_factor = col_factor, col_pars = col_pars,col_valor = col_valor)
}
}
