% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources_fn_abiotic.R
\name{fn_plot_bar_abiotic}
\alias{fn_plot_bar_abiotic}
\title{fn_plot_bar_abiotic}
\usage{
fn_plot_bar_abiotic(
  data,
  col_pars,
  col_sitio,
  col_valor,
  col_factor = NULL,
  col_unidad,
  code_sitio,
  matriz,
  data_pars,
  ord_sitio = "asc",
  width = 8,
  height = 6,
  ord_factor = NULL,
  aspect_ratio = 1
)
}
\arguments{
\item{data}{archivo entrada que tiene que tener, al menos, las columnas: sitio, parámetros (nombre o sigla), valor de parámetros. Tiene que estar en formato "long". Se recomida usar csv o tsv como formatos.}

\item{col_pars}{string que indica el nombre de la columna en data que tiene los parámetros (su sigla).}

\item{col_sitio}{string que indica el nombre de la columna que tiene los sitios de muestreo.}

\item{col_valor}{string que indica el nombre de la columna que tiene el valor de los parámetros.}

\item{col_factor}{string que indica el nombre del de la columna que contiene una variable de agrupamiento para el plot (e.g. zonas, campañas, etc.). Por defecto, Nulo.}

\item{code_sitio}{string que indica el código (generalmente una letra seguida un guión) que utilizada para nombras las estaciones (eg. "E-", "P-", "D-"). Se esperan puntos de muestreos asignados de forma consecutiva y sin gaps.}

\item{matriz}{string que indica el nombre de la "matriz" de datos que se esta analizando. Sirve principalmente como ID para las salidas y para diferenciar entre sedimentos y aguas.}

\item{data_pars}{"dataframe" que contenga los parámetros (sus siglas) tal como aparecen en el "dataframe" de entrada, junto a una columna que indique a que tipo de parámetro pertenece (i.e. Físico, Químico, Nutriente, Metal, Orgánico e inorgánico y Biológico).Las columnas se tienen que llamar: Param y cat_pars}

\item{ord_sitio}{string que indica el orden en que se quiere que aparezcan los sitios en el gráfico (de forma ascendente o descendente de acuerdo al número de estación o punto de muestreo). Por defecto, toma valor "asc" (órden ascendente)}

\item{width}{ancho del gráfico. Por defecto, toma valor 8}

\item{height}{alto del gráfico. Por defecto, toma valor 6}

\item{ord_factor}{string que indica el orden en el que se quiera que aparezan en las leyendas y/o ejes los iteml del grupo. Solo tiene sentido si se usa col_grupo. Por defecto, Nulo.}

\item{aspect_ratio}{relación alto-ancho. Por defecto, toma valor 1.}
}
\value{
gráfico de barras por tipo de parámetros
}
\description{
función para graficar las concentraciones o niveles de los parámetros a los largo de las estaciones de muestro, coloreadolos por tipo (físicos, químicos, nutrientes, metales, orgánico e inorgánico y biológico; en este mismo orden)
}
\examples{
\dontrun{
Plot sin grupos
data <- readr::read_tsv("data_agua_RDLP.tsv")
col_pars <- "Param"
col_sitio <- "Sitio"
col_valor <- "Resultado"
col_unidad <- "Unidad"
matriz <- "agua"
code_sitio <- "P-"
ord_sitio <- "asc"# "desc o asc"
data_pars <- readr::read_tsv("tabla_pars_master.tsv")
width <- 8
height <- 11
fn_plot_bar_abiotic(data = data,col_pars = col_pars,col_sitio = col_sitio,col_valor = col_valor,aspect_ratio = 1,col_unidad = col_unidad,code_sitio = code_sitio,matriz = matriz,data_pars = data_pars,ord_sitio = ord_sitio,width = width,height = height)
# con 1 grupo
data <- readr::read_tsv("data_agua_VEN.tsv")
col_pars <- "Param"
col_sitio <- "Sitio"
col_valor <- "Valor"
col_unidad <- "Unidad"
matriz <- "agua"
code_sitio <- "P-"
ord_sitio <- "asc"# "desc o asc"
data_pars <- readr::read_tsv("tabla_pars_master.tsv")
col_factor <- "zonas"
width <- 9
height <- 8
fn_plot_bar_abiotic(data = data,col_pars = col_pars,col_sitio = col_sitio,col_valor = col_valor,col_factor = col_factor,aspect_ratio = 1,col_unidad = col_unidad,code_sitio = code_sitio,matriz = matriz,data_pars = data_pars,ord_sitio = ord_sitio,width = width,height = height)
}
}
