% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources_fn_biotic.R
\name{fn_plot_bar_biotic}
\alias{fn_plot_bar_biotic}
\title{fn_plot_bar_biotic}
\usage{
fn_plot_bar_biotic(
  data,
  col_sitio,
  col_N,
  col_zonas = NULL,
  col_taxa,
  taxa_grupo,
  ord_sitio = "asc",
  ord_zonas = NULL,
  taxa_id,
  code_sitio,
  width = 12,
  height = 12
)
}
\arguments{
\item{data}{archivo entrada que tiene que tener, al menos, las columnas que representen al taxón, grupo taxonómico, sitios y N. Tiene que estar en formato "long". Se recomida usar csv o tsv como formatos.}

\item{col_sitio}{cadena de texto que indica el nombre de la columna que tiene los sitios de muestreo.}

\item{col_N}{cadena de texto que indica el nombre de la columna que contiene el N (abundancia) en la base de datos de entrada.}

\item{col_taxa}{cadena de texto que indica el nombre de la columna que contiene a los taxones (generalmente se llama especie, taxa, etc.) en la base de datos de entrada.}

\item{taxa_grupo}{cadena de texto que indica la clasificación taxonómica de los taxones (e.g. Familia, Clase, etc.) en la base de datos de entrada.}

\item{ord_sitio}{cadena de texto que indica el orden en el que se quiere aparezcan los sitios en el gráfico (de forma ascendente o descendente de acuerdo al número de estaciones o puntos de muestreo). Por defecto, toma valor "asc" (orden ascendente).}

\item{ord_zonas}{Si aplica, cadena de texto que indica el orden (en la leyenda, ejes, etc.) del grupo o factor de agrupamiento. Por defecto, NULO.}

\item{taxa_id}{cadena de texto que indica el nombre del "grupo funcional" al cual pertenecen los taxones ("fitoplancton", "zooplancton", "macrofitas", "perifiton", "ictiofauna",etc.)}

\item{code_sitio}{cadena de texto que indica el código (generalmente una letra seguida un guión) que utilizada para nombras las estaciones (eg. "E-", "P-", "D-").}

\item{width}{ancho del gráfico. Por defecto, toma valor 12}

\item{height}{alto del gráfico. Por defecto, toma valor 12}

\item{col_zona}{Si aplica, cadena de texto que indica el nombre de la columna que contiene los grupos o factores de agrupamiento en la base de datos de entrada.}
}
\value{
gráfico de barras de abundancia/riqueza de especies de taxones por grupo taxonómico
}
\description{
función para graficar de barras para la abundancia de las especies/taxones de un taxón determinado
}
\examples{
\dontrun{
# Para gráfico sin variable o factor de agrupamiento 
data <- read_tsv("data_fitoplancton.tsv")
col_taxa <- "Taxa"
taxa_grupo <- "Clase"
col_N <- "N"
col_sitio <- "Sitio"
ord_sitio <- "asc"
code_sitio <- "P-"
taxa_id <- "fitoplancton"
width <- 12
height <- 12
fn_plot_bar_biotic(data = data, col_taxa = col_taxa, taxa_grupo = taxa_grupo,col_N = col_N, col_sitio = col_sitio, ord_sitio = ord_sitio, code_sitio = code_sitio, taxa_id = taxa_id, width = width, height = height)
# Para gráfico con 1 variable o factor de agrupamiento, en orden alfanumérico. 
data <- read_tsv("data_fitoplancton.tsv")
#se necesita que la variable o factor este en el dataframe
data_grupos <- read_tsv("data_grupos.tsv") 
data <- left_join(data, data_grupos, by = c("Sitio" = "Sitio"))
col_taxa <- "Taxa"
taxa_grupo <- "Clase"
col_N <- "N"
col_sitio <- "Sitio"
ord_sitio <- "asc"
code_sitio <- "P-"
col_zonas <- "Grupo"
taxa_id <- "fitoplancton"
width <- 12
height <- 12
fn_plot_bar_biotic(data = data, col_taxa = col_taxa, taxa_grupo = taxa_grupo,col_N = col_N, col_sitio = col_sitio, ord_sitio = ord_sitio, code_sitio = code_sitio, col_zonas = col_zonas, taxa_id = taxa_id, width = width, height = height)
# Para gráfico con 1 variable o factor de agrupamiento, en algún orden definido. 
data <- read_tsv("data_fitoplancton.tsv")
#se necesita que la variable o factor este en el dataframe
data_grupos <- read_tsv("data_grupos.tsv") 
data <- left_join(data, data_grupos, by = c("Sitio" = "Sitio"))
col_taxa <- "Taxa"
taxa_grupo <- "Clase"
col_N <- "N"
col_sitio <- "Sitio"
ord_sitio <- "asc"
code_sitio <- "P-"
col_zonas <- "Grupo"
ord_zonas <- c("Q1", "Q2", "Q3", "Q4", "Q5","L")
taxa_id <- "fitoplancton"
width <- 12
height <- 12
fn_plot_bar_biotic(data = data, col_taxa = col_taxa, taxa_grupo = taxa_grupo,col_N = col_N, col_sitio = col_sitio, ord_sitio = ord_sitio, code_sitio = code_sitio, col_zonas = col_zonas,ord_zonas = ord_zonas, taxa_id = taxa_id, width = width, height = height)
}
}
