% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources_fn_abiotic.R
\name{fn_plot_correlogram}
\alias{fn_plot_correlogram}
\title{fn_plot_correlogram}
\usage{
fn_plot_correlogram(
  data,
  col_pars,
  col_sitio,
  col_factor = NULL,
  matriz,
  code_sitio,
  data_pars,
  width = 6,
  height = 6
)
}
\arguments{
\item{data}{archivo entrada que tiene que tener, al menos, las columnas: sitio, parámetros (nombre o sigla), valor de parámetros. Tiene que estar en formato "long". Se recomida usar csv o tsv como formatos.}

\item{col_pars}{string que indica el nombre de la columna en data que tiene los parámetros (su sigla).}

\item{col_sitio}{string que indica el nombre de la columna que tiene los sitios de muestreo.}

\item{matriz}{string que indica el nombre de la "matriz" de datos que se esta analizando. Sirve principalmente como ID para las salidas y para diferenciar entre sedimentos y aguas.}

\item{code_sitio}{string que indica el código (generalmente una letra seguida un guión) que utilizada para nombras las estaciones (eg. "E-", "P-", "D-"). Se esperan puntos de muestreos asignados de forma consecutiva y sin gaps.}

\item{data_pars}{"dataframe" que contenga los parámetros (sus siglas) tal como aparecen en el "dataframe" de entrada, junto a un columna que indique a que tipo de parámetro pertenece (i.e. Físico, Químico, Nutriente, Metal, Orgánico e inorgánico y Biológico).Las columnas se tienen que llamar: Param y cat_pars}

\item{width}{ancho del gráfico. Por defecto, toma valor 6}

\item{height}{alto del gráfico. Por defecto, toma valor 6}

\item{col_valor}{string que indica el nombre de la columna que tiene el valor de los parámetros.}
}
\value{
gráfico de correlaciones de spearmann
}
\description{
función para graficar el correlograma de las variables.
}
\examples{
\dontrun{
data <- readr::read_tsv("data_sedimento (copia).tsv")
col_pars <- "Param"
col_sitio <- "Estacion"
col_valor <- "Valor"
matriz <- "sedimento"
code_sitio <- "E-"
data_pars <- readr::read_tsv("tabla_pars.tsv")
width <- 8
height <- 9
fn_plot_correlogram(data = data, col_pars = col_pars, col_sitio = col_sitio, matriz = matriz, code_sitio, code_sitio, data_pars = data_pars, width = width,height = width)
}
}
