% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources_fn_abiotic.R
\name{fn_plot_granulometria}
\alias{fn_plot_granulometria}
\title{fn_plot_granulometria}
\usage{
fn_plot_granulometria(
  data,
  col_pars,
  col_sitio,
  sitios_custom = NULL,
  col_valor,
  code_sitio,
  ord_sitio = "asc",
  col_grupo = NULL,
  ord_factor = NULL,
  width = 8,
  height = 6,
  aspect_ratio = 1
)
}
\arguments{
\item{data}{archivo entrada que tiene que tener, al menos, las columnas: sitio, parámetros (nombre o sigla), valor de parámetros. Tiene que estar en formato "long". Se recomida usar csv o tsv como formatos.}

\item{col_pars}{string que indica el nombre de la columna en data que tiene los parámetros (su sigla).}

\item{col_sitio}{string que indica el nombre de la columna que tiene los sitios de muestreo.}

\item{sitios_custom}{string con las estaciones en un orden en particular, solo aplica si ord_sitio es "custom"}

\item{col_valor}{string que indica el nombre de la columna que tiene el valor de los parámetros.}

\item{code_sitio}{string que indica el código (generalmente una letra seguida un guión) que utilizada para nombras las estaciones (eg. "E-", "P-", "D-"). Se esperan puntos de muestreos asignados de forma consecutiva y sin gaps.}

\item{ord_sitio}{string que indica el orden en que se quiere que aparezcan los sitios en el gráfico (de forma ascendente o descendente o custom).}

\item{col_grupo}{string que indica el nombre del de la columna que contiene una variable de agrupamiento para el plot (e.g. zonas, campañas, etc.). Por defecto, Nulo.}

\item{ord_factor}{string que indica el orden en el que se quiera que aparezan en las leyendas y/o ejes los iteml del grupo. Solo tiene sentido si se usa col_grupo. Por defecto, Nulo.}

\item{width}{ancho del gráfico. Por defecto, toma valor 8}

\item{height}{alto del gráfico. Por defecto, toma valor 6}

\item{aspect_ratio}{relación alto-ancho. Por defecto, toma valor 1.}
}
\value{
gráfico de granulometría
}
\description{
función para graficar fracciones granulométricas. Solo tiene sentido para datos de Sedimentos.
}
\examples{
\dontrun{
# figura sin grupos
data <- readr::read_tsv("data_sedimento_RDLP.tsv")
col_pars <- "Param"
col_sitio <- "Estacion"
col_valor <- "Resultado"
code_sitio <- "P-"
ord_sitio <- "asc"# "desc o asc"
width <- 6
height <- 4
fn_plot_granulometria(data = data,col_pars = col_pars,col_sitio = col_sitio, col_valor = col_valor,ord_sitio = ord_sitio, width = 8,height = 6, aspect_ratio = 1, code_sitio = code_sitio)
}
}
