% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources_fn_biotic.R
\name{fn_plot_heat_pres}
\alias{fn_plot_heat_pres}
\title{fn_plot_heat_pres}
\usage{
fn_plot_heat_pres(
  data,
  col_taxa,
  col_N,
  col_facet = NULL,
  ord_facet = NULL,
  col_zonas = NULL,
  ord_zonas = NULL,
  col_sitio = NULL,
  ord_sitio = NULL
)
}
\value{
heatmap de presencia
}
\description{
función para graficar presencia ausencia de taxa
}
