% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources_fn_biotic.R
\name{fn_plot_pie}
\alias{fn_plot_pie}
\title{fn_plot_pie}
\usage{
fn_plot_pie(
  data,
  taxa_grupo,
  col_N,
  cumsum_cut = 90,
  n_size = 5,
  width = 12,
  height = 13
)
}
\arguments{
\item{data}{archivo entrada que tiene que tener, al menos, las columnas que representen al taxón, grupo taxonómico, taxones y N. Tiene que estar en formato "long". Se recomida usar csv o tsv como formatos.}

\item{taxa_grupo}{cadena de texto que indica la columna que contiene la clasificación taxonómica de los taxones (e.g. Familia, Clase, etc) en la base de datos de entrada.}

\item{col_N}{cadena de texto que indica el nombre de la columna que contiene el N (abundancia) en la base de datos de entrada.}

\item{cumsum_cut}{Valor entero o decimal de la suma acumulada hasta la cual se quiera representar sin la etiqueta de "Otros". Por defecto, toma valor 90.}

\item{n_size}{Valor entero que indica el número de piezas en el gráfico de torta. Por defecto, toma valor de 5.}

\item{width}{ancho del gráfico. Por defecto, toma valor 12.}

\item{height}{alto del gráfico. Por defecto, toma valor 13.}

\item{col_taxa}{cadena de texto que indica el nombre de la columna que contiene a los taxones (generalmente se llama especie, taxa, etc) en la base de datos de entrada.}
}
\value{
gráfico de barras de abundancia/riqueza de especies de taxones por grupo taxonómico
}
\description{
función para graficar la composición porcentual de taxones por grupos taxonómicos en gráficos de tortas.
}
\examples{
\dontrun{
# Para gráfico sin variable o factor de agrupamiento
data <- read_tsv("data_fitoplancton.tsv")
taxa_id <- "fitoplancton"
col_N <- "N"
col_taxa <- "Taxa"
taxa_grupo <- "Clase"
cumsum_cut <- 90
n_size <- 5
width = width
height = height
fn_plot_pie(data = data, taxa_grupo = taxa_grupo, cumsum_cut = 90, n_size = 5, width = width, height = height)
}
}
