% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources_fn_biotic.R
\name{fn_plot_spec_rich}
\alias{fn_plot_spec_rich}
\title{fn_plot_spec_rich}
\usage{
fn_plot_spec_rich(
  data,
  taxa_group,
  col_taxa,
  col_N,
  col_facet,
  ord_facet = NULL,
  col_zonas = NULL,
  ord_zonas = NULL,
  col_sitio = NULL,
  ord_sitio = NULL,
  taxa_id,
  height = 6,
  width = 8
)
}
\arguments{
\item{data}{archivo entrada que tiene que tener, al menos, las columnas que contengan a los taxones, N y sitios de muestreo. Tiene que estar en formato "long". Se recomida usar csv o tsv como formatos.}

\item{col_taxa}{cadena de texto que indica el nombre de la columna que contiene los taxones (generalmente se llama especie, taxa, sigla) en la base de datos de entrada.}

\item{col_N}{cadena de texto que indica el nombre de la columna que contiene el N (abundancia) en la base de datos de entrada.}

\item{col_facet}{cadena de texto que indica el nombre de la columna que indica la variable para hacer el facet. Si no se proporciona orden , asume el orden  de la variable en el dataframae}

\item{col_sitio}{cadena de texto que indica el nombre de la columna que contiene los sitios de muestreo en la base de datos de entrada. Puede ser null}

\item{taxa_id}{cadena de texto que indica el nombre del "grupo funcional" al cual pertenecen los taxones ("fitoplancton", "zooplancton", "macrofitas", "perifiton", "ictiofauna",etc.).}

\item{height}{alto del gráfico. Por defecto, toma valor 6.}

\item{width}{ancho del gráfico. Por defecto, toma valor 8.}

\item{col_zona}{cadena de texto que indica el nombre de la columna que contiene al factor de agrupamiento (zonas, campañas, etc) en la base de datos de entrada. Por defecto, Nulo.}
}
\value{
gráfico de barras
}
\description{
función para graficar riqueza de taxones por taxa por sitios o zonas en grafico de barra
}
\examples{
\dontrun{
data <- read_tsv("data_fitoplancton_integrado.tsv")
col_N <- "N"
col_facet <- "Campaña"
col_sitio <- "Sitio"
ord_sitio <- paste0("P-", seq(1,6))
ord_facet <- c("Verano 2024", "Invierno 2024")
taxa_id <- "fitoplancton"
taxa_group <- "Clase"
col_taxa <- "Taxa"
width <- 8
height <- 8
fn_plot_spec_rich(data = data, taxa_group = taxa_group,col_taxa = col_taxa,col_N = col_N,col_facet = col_facet,ord_facet = ord_facet,col_sitio = col_sitio,ord_sitio = ord_sitio,taxa_id = taxa_id,height = height, width = width)
}
}
