% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources_fn_biotic.R
\name{fn_plot_treemap}
\alias{fn_plot_treemap}
\title{fn_plot_treemap}
\usage{
fn_plot_treemap(
  data,
  col_sitio = NULL,
  col_N,
  col_facet = NULL,
  ord_facet = NULL,
  col_zonas = NULL,
  taxa_id,
  taxa_group,
  width = 8,
  height = 10
)
}
\arguments{
\item{data}{archivo entrada que tiene que tener, al menos, las columnas que contengan a los taxones, N y sitios de muestreo. Tiene que estar en formato "long". Se recomida usar csv o tsv como formatos.}

\item{col_sitio}{cadena de texto que indica el nombre de la columna que contiene los sitios de muestreo en la base de datos de entrada.}

\item{col_N}{cadena de texto que indica el nombre de la columna que contiene el N (abundancia) en la base de datos de entrada.}

\item{col_facet}{cadena de texto que indica el nombre de la columna que indica la variable para hacer el facet. Si no se proporciona orden , asume el orden  de la variable en el dataframae}

\item{taxa_id}{cadena de texto que indica el nombre del "grupo funcional" al cual pertenecen los taxones ("fitoplancton", "zooplancton", "macrofitas", "perifiton", "ictiofauna",etc.).}

\item{width}{ancho del gráfico. Por defecto, toma valor 8.}

\item{height}{alto del gráfico. Por defecto, toma valor 10.}

\item{col_taxa}{cadena de texto que indica el nombre de la columna que contiene los taxones (generalmente se llama especie, taxa, sigla) en la base de datos de entrada.}

\item{col_zona}{cadena de texto que indica el nombre de la columna que contiene al factor de agrupamiento (zonas, campañas, etc) en la base de datos de entrada. Por defecto, Nulo.}
}
\value{
gráfico treemap.
}
\description{
función para graficar abundancia de taxones por taxa por sitios o zonas en forma de treemap
}
\examples{
\dontrun{
data <- read_tsv("data_fitoplancton_integrado.tsv")
col_N <- "N"
col_facet <- "Campaña"
col_sitio <- "Sitio"
ord_facet <- c("Verano 2024", "Invierno 2024")
taxa_id <- "fitoplancton"
taxa_group <- "Clase"
width <- 8
height <- 10
fn_plot_treemap(data = data,col_facet = col_facet,ord_facet = ord_facet,col_sitio = col_sitio,col_N = col_N,taxa_id = taxa_id,taxa_group = taxa_group,width = width,height = height)
}
}
