% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources_fn_abiotic.R
\name{fn_stats}
\alias{fn_stats}
\title{fn_stat}
\usage{
fn_stats(data, col_pars, col_valor, data_pars, matriz, round = 2)
}
\arguments{
\item{data}{archivo de entrada. Tiene que tener, al menos, las columnas: sitio, parámetros (nombre o sigla), valor de parámetros. Tiene que estar en formato "long". Se recomida usar csv o tsv como formatos.}

\item{col_pars}{string que indica el nombre de la columna en data que tiene los parámetros (su sigla).}

\item{col_valor}{string que indica el nombre de la columna que tiene el valor de los parámetros.}

\item{data_pars}{"dataframe" que contenga los parámetros (sus siglas) tal como aparecen en el "dataframe" de entrada, junto a un columna que indique a que tipo de parámetro pertenece (i.e. Físico, Químico, Nutriente, Metal, Orgánico e inorgánico y Biológico).Las columnas se tienen que llamar: Param y cat_pars}

\item{matriz}{string que indica el nombre de la "matriz" de datos que se esta analizando. Sirve principalmente como ID para las salidas y para diferenciar entre sedimentos y aguas.}

\item{round}{integro de determina el número de decimales para redondeo en la tabla. Por defecto, round =2.}
}
\value{
tabla básica estadisticos
}
\description{
función para obtener la tabla de estadística descriptiva.
}
\examples{
\dontrun{
data <- readr::read_tsv("data_agua_RDLP_RDLP.tsv")
col_pars <- "Param"
data_pars <- readr::read_tsv("tabla_pars_master.tsv")
col_valor <- "Resultado"
matriz <- "agua"
round <- 4
fn_stats(data = data,col_pars = col_pars,col_valor = col_valor,data_pars = data_pars,matriz = matriz,round = round)
}
}
