% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_data_pca_plot}
\alias{get_data_pca_plot}
\title{Calcular coordenadas para Biplot PCA (Función Interna)}
\usage{
get_data_pca_plot(pca_obj)
}
\description{
Calcular coordenadas para Biplot PCA (Función Interna)
}
\keyword{internal}
