% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources_fn_biotic.R
\name{get_index_df}
\alias{get_index_df}
\title{Calcular Índices de Diversidad (N, S, H)}
\usage{
get_index_df(
  data,
  col_value,
  col_site,
  col_sampling = NULL,
  col_taxa,
  col_zone = NULL,
  col_rep = NULL,
  output_name,
  output_dir = "outputs",
  fun = sum
)
}
\arguments{
\item{data}{Data frame con los datos.}

\item{col_value}{String. Nombre de la columna de abundancia o valor.}

\item{col_site}{String. Nombre de la columna de sitios.}

\item{col_sampling}{String (Opcional). Nombre de la columna de muestreo.}

\item{col_taxa}{String. Nombre de la columna de taxones.}

\item{col_zone}{String (Opcional). Nombre de la columna de zonas.}

\item{col_rep}{String (Opcional). Nombre de la columna de réplicas.}

\item{output_name}{String. Nombre del archivo de salida (.csv).}

\item{output_dir}{String. Directorio de salida. Default "outputs".}

\item{fun}{Función o nombre de función para agregar datos. Default sum.}
}
\value{
Un data frame con los índices calculados (invisible).
}
\description{
Agrupa datos, calcula abundancia (N), riqueza (S) y Shannon (H)
  usando \code{vegan}, y guarda el resultado en CSV.
}
