% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources_fn_biotic.R
\name{get_taxo_worms}
\alias{get_taxo_worms}
\title{Taxonomía según WoRMS}
\usage{
get_taxo_worms(taxa_vector, max_dist = 3, dist_method = "lv")
}
\arguments{
\item{taxa_vector}{Vector de caracteres con los nombres de las especies.}

\item{max_dist}{Distancia máxima para match difuso. Default 3.}

\item{dist_method}{Método de cálculo de distancia. Default "lv".}
}
\value{
Un tibble con la jerarquía taxonómica y metadatos de búsqueda.
}
\description{
Toma un vector de nombres científicos y consulta la base de
  datos de WoRMS para obtener su clasificación taxonómica completa.
}
