% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources_fn_abiotic.R
\name{plot_granulometria}
\alias{plot_granulometria}
\title{Gráfico de Distribución Granulométrica (Stacked Barplot)}
\usage{
plot_granulometria(
  data,
  col_pars,
  col_site,
  col_value,
  col_factor = NULL,
  col_rep = NULL,
  ord_site = NULL,
  levels_factor = NULL,
  width = 8,
  height = 6,
  output_name = "plot_granulometria.png",
  output_dir = "outputs",
  save_rds = FALSE
)
}
\arguments{
\item{data}{Data frame con los datos.}

\item{col_pars}{String. Columna con los tipos de grano.}

\item{col_site}{String. Columna con los sitios.}

\item{col_value}{String. Columna con el porcentaje/valor.}

\item{col_factor}{String (Opcional). Columna para facetas (columnas).}

\item{col_rep}{String (Opcional). Columna para facetas (filas).}

\item{ord_site}{Vector (Opcional). Orden de sitios.}

\item{levels_factor}{Vector (Opcional). Orden del factor.}

\item{width}{Numérico. Ancho.}

\item{height}{Numérico. Alto.}

\item{output_name}{String. Nombre archivo salida.}

\item{output_dir}{Ruta del directorio de salida. Default: \code{"outputs"}.}

\item{save_rds}{Lógico. Guardar RDS.}
}
\value{
Un objeto ggplot.
}
\description{
Barras apiladas para composición granulométrica por sitio.
}
