% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources_fn_biotic.R
\name{plot_index}
\alias{plot_index}
\title{Graficar Variación Espacial de Índices Comunitarios}
\usage{
plot_index(
  data,
  col_value,
  col_index,
  col_site,
  ord_site = NULL,
  col_rep = NULL,
  col_facet = NULL,
  ord_facet = NULL,
  taxa_id,
  legend_name,
  width = 8,
  height = 6,
  output_name = "plot_index.png",
  xlabel = "Sitios",
  save_rds = FALSE
)
}
\arguments{
\item{data}{Data frame con los datos.}

\item{col_value}{String. Nombre de la columna con el valor del índice.}

\item{col_index}{String. Nombre de la columna que indica el tipo de índice (N, S, H).}

\item{col_site}{String. Nombre de la columna de sitios (Eje X).}

\item{ord_site}{Vector (Opcional). Orden de los niveles de sitios.}

\item{col_rep}{String (Opcional). Nombre de la columna para agrupar líneas/colores (ej. Réplica, Campaña, Estrato).}

\item{col_facet}{String (Opcional). Nombre de la columna para facetar horizontalmente (ej. Zona).}

\item{ord_facet}{Vector (Opcional). Orden de los niveles del facet.}

\item{taxa_id}{String. Identificador del taxón para el título.}

\item{legend_name}{String. Título de la leyenda. Determina la paleta de colores ("estratos" activa paleta azul fija).}

\item{width}{Numérico. Ancho de la imagen. Default 8.}

\item{height}{Numérico. Alto de la imagen. Default 6.}

\item{output_name}{String. Nombre del archivo de salida.}

\item{xlabel}{String. Etiqueta para el eje X. Default "Sitios".}

\item{save_rds}{Lógico. Si es TRUE, guarda también el objeto R (.rds). Por defecto FALSE.}
}
\value{
Un objeto ggplot (invisible).
}
\description{
Crea un gráfico de líneas y puntos facetado por tipo de índice (N, S, H). Asigna paletas de colores específicas según el nombre de la leyenda.
}
