% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources_fn_biotic.R
\name{plot_index_by_taxa}
\alias{plot_index_by_taxa}
\title{Graficar Índices de Diversidad por Taxón}
\usage{
plot_index_by_taxa(
  data,
  col_site,
  ord_site = NULL,
  col_N,
  col_taxa,
  taxon_group,
  col_facet = NULL,
  ord_facet = NULL,
  taxa_id,
  width = 10,
  height = 6,
  xlabel = NULL
)
}
\arguments{
\item{data}{Data frame con los datos.}

\item{col_site}{String. Nombre de la columna de sitios.}

\item{ord_site}{Vector (Opcional). Orden de los niveles de sitios.}

\item{col_N}{String. Nombre de la columna de abundancia.}

\item{col_taxa}{String. Nombre de la columna de taxones individuales (para conteo de riqueza).}

\item{taxon_group}{String. Nombre de la columna de agrupación taxonómica (ej. "Phylum", "Clase") para el relleno del gráfico.}

\item{col_facet}{String (Opcional). Nombre de la columna para facetar (ej. "Zona").}

\item{ord_facet}{Vector (Opcional). Orden de los niveles del facet.}

\item{taxa_id}{String. Identificador del grupo taxonómico general (ej. "Macroinvertebrados") para títulos y nombres de archivo.}

\item{width}{Numérico. Ancho de la imagen.}

\item{height}{Numérico. Alto de la imagen.}

\item{xlabel}{String (Opcional). Etiqueta personalizada para el eje X. Por defecto "Sitios".}
}
\value{
Invisible NULL. Guarda dos archivos PNG.
}
\description{
Genera y guarda dos gráficos de barras apiladas: uno de Abundancia Total y otro de Riqueza de Taxones, desglosados por sitio y grupo taxonómico.
}
