% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources_fn_biotic.R
\name{plot_nmds}
\alias{plot_nmds}
\title{Plot NMDS and PERMANOVA}
\usage{
plot_nmds(
  data,
  col_site,
  col_N,
  col_taxa,
  taxa_id,
  col_factor = NULL,
  legend_factor = NULL,
  levels_factor = NULL,
  height = 6,
  width = 6,
  output_name = "plot_nmds.png",
  save_rds = FALSE
)
}
\arguments{
\item{data}{Data frame con los datos.}

\item{col_site}{String. Nombre de la columna de sitios.}

\item{col_N}{String. Nombre de la columna de abundancia.}

\item{col_taxa}{String. Nombre de la columna de taxones.}

\item{taxa_id}{String. Nombre del grupo taxonómico para el título (ej. "Fitoplancton").}

\item{col_factor}{String (Opcional). Nombre de la columna de agrupación. Si es NULL, se usa `col_site` para colorear.}

\item{legend_factor}{String (Opcional). Título personalizado para la leyenda del gráfico cuando se usa `col_factor`.}

\item{levels_factor}{Vector (Opcional). Orden de los niveles del factor.}

\item{height}{Numérico. Alto de la imagen. Default 6.}

\item{width}{Numérico. Ancho de la imagen. Default 6.}

\item{output_name}{String. Nombre del archivo de salida.}

\item{save_rds}{Lógico. Si es TRUE, guarda el objeto ggplot como .rds.}
}
\value{
Un objeto ggplot con el NMDS.
}
\description{
Realiza un análisis NMDS con transformación de Hellinger. Si hay factor, dibuja elipses y calcula PERMANOVA. Si no, colorea por sitios.
}
