% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources_fn_abiotic.R
\name{plot_parameter_levels}
\alias{plot_parameter_levels}
\title{Gráficos de barras de parámetros abióticos con normas}
\usage{
plot_parameter_levels(
  data,
  col_pars,
  col_site,
  ord_site,
  col_value,
  col_facet = NULL,
  ord_facet = NULL,
  col_units,
  matrix,
  type_par,
  col_group = NULL,
  legend_name = NULL,
  data_source,
  ref_data = NULL
)
}
\arguments{
\item{data}{Tabla de datos de entrada.}

\item{col_pars}{Nombre de la columna (cadena) que identifica el parámetro.}

\item{col_site}{Nombre de la columna (cadena) que identifica el sitio.}

\item{ord_site}{Vector con el orden deseado de los sitios en el eje x.}

\item{col_value}{Nombre de la columna (cadena) con los valores numéricos.}

\item{col_facet}{Nombre de la columna (cadena) usada para facetear (opcional).}

\item{ord_facet}{Vector con el orden deseado de los niveles de la faceta (opcional).}

\item{col_units}{Nombre de la columna (cadena) con la unidad del parámetro.}

\item{matrix}{Cadena que describe la matriz (por ejemplo, "agua", "sedimento").}

\item{type_par}{Nombre de la columna (cadena) que clasifica los parámetros por tipo.}

\item{col_group}{Nombre de la columna (cadena) con el grupo (estrato, réplica, etc.) para colorear barras (opcional). En caso de ser estratos, tienen que ser ,"Supreficie", "Medio" y/o "Fondo".}

\item{legend_name}{Título de la leyenda asociada a `col_group` (opcional).}

\item{data_source}{Cadena que describe la fuente de datos (por ejemplo, "in situ").}

\item{ref_data}{Tabla con normas o valores de referencia para los parámetros (incluyendo columnas `Sigla`, `nombre_norma`, `limite`, `tipo_limite`).}
}
\description{
Variante de `plot_parameter_levels` que, cuando se entrega una tabla de normas (`ref_data`), filtra y dibuja solo aquellas referencias que se sexcedan
}
