% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources_fn_abiotic.R
\name{plot_pca}
\alias{plot_pca}
\title{Biplot de PCA con PERMANOVA Opcional}
\usage{
plot_pca(
  data,
  col_pars,
  col_site,
  col_value,
  type_par,
  matrix,
  col_rep = NULL,
  col_factor = NULL,
  levels_factor = NULL,
  width = 6,
  height = 6,
  dist = "euclidean",
  output_name = "pca.png",
  save_rds = FALSE
)
}
\arguments{
\item{data}{Data frame con los datos.}

\item{col_pars}{String. Columna con los nombres de los parámetros.}

\item{col_site}{String. Columna con los identificadores de sitio.}

\item{col_value}{String. Columna con los valores numéricos.}

\item{type_par}{String. Columna que clasifica el tipo de parámetro.}

\item{matrix}{String. Tipo de matrix (ej. "Agua", "Sedimento").}

\item{col_rep}{String (Opcional). Columna de réplicas.}

\item{col_factor}{String (Opcional). Columna para agrupar y calcular PERMANOVA.}

\item{levels_factor}{Vector (Opcional). Orden específico para los niveles de `col_factor`.}

\item{width}{Numérico. Ancho de la imagen.}

\item{height}{Numérico. Alto de la imagen.}

\item{dist}{String. Método de distancia para PERMANOVA (por defecto "euclidean").}

\item{output_name}{String. Nombre del archivo de salida (png).}

\item{save_rds}{Lógico. Si es TRUE, guarda también el objeto R (.rds). Por defecto FALSE.}
}
\value{
Un objeto ggplot (invisible).
}
\description{
Realiza un Análisis de Componentes Principales (PCA) y genera un biplot.
Si se proporciona un factor (`col_factor`), calcula PERMANOVA y añade elipses.
}
