% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils_get_pca_biplot}
\alias{utils_get_pca_biplot}
\title{Calcular coordenadas para Biplot PCA (Función Interna)}
\usage{
utils_get_pca_biplot(pca_obj)
}
\description{
Calcular coordenadas para Biplot PCA (Función Interna)
}
\keyword{internal}
