% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources_fn_abiotic.R
\name{fn_plot_granulometria}
\alias{fn_plot_granulometria}
\title{fn_plot_granulometria}
\usage{
fn_plot_granulometria(
  data,
  col_pars,
  col_sitio,
  col_valor,
  code_sitio,
  ord_sitio = "asc",
  col_grupo = NULL,
  ord_grupo = NULL,
  width,
  height,
  aspect_ratio = 1
)
}
\arguments{
\item{data}{archivo entrada que tiene que tener, al menos, las columnas: sitio, parámetros (nombre o sigla), valor de parámetros. Tiene que estar en formato "long". Se recomida usar csv o tsv como formatos.}

\item{col_pars}{string que indica el nombre de la columna en data que tiene los parámetros (su sigla).}

\item{col_sitio}{string que indica el nombre de la columna que tiene los sitios de muestreo.}

\item{col_valor}{string que indica el nombre de la columna que tiene el valor de los parámetros.}

\item{code_sitio}{string que indica el código (generalmente una letra seguida un guión) que utilizada para nombras las estaciones (eg. "E-", "P-", "D-"). Se esperan puntos de muestreos asignados de forma consecutiva y sin gaps.}

\item{ord_sitio}{string que indica el orden en que se quiere que aparezcan los sitios en el gráfico (de forma ascendente o descendente de acuerdo al número de estación o punto de muestreo).}

\item{col_grupo}{string que indica el nombre del de la columna que contiene una variable de agrupamiento para el plot (e.g. zonas, campañas, etc.)}

\item{ord_grupo}{string que indica el orden en el que se quiera que aparezan en las leyendas y/o ejes los iteml del grupo. Solo tiene sentido si se usa col_grupo}
}
\value{
gráfico granulometría
}
\description{
función para graficar fracciones granulométricas. Solo tiene sentido para datos de Sedimentos.
}
