% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources_fn_abiotic.R
\name{get_glm}
\alias{get_glm}
\title{Modelos Lineales Generalizados (GLM) por Grupos}
\usage{
get_glm(
  data,
  col_pars,
  col_value,
  col_factor,
  interaction = NULL,
  output_name = NULL,
  family = c("Gaussian", "Gamma", "Quasipoisson"),
  dec = 2
)
}
\arguments{
\item{data}{Data frame con los datos.}

\item{col_pars}{String. Nombre de la columna que identifica el parámetro (variable de agrupación).}

\item{col_value}{String. Nombre de la columna con la variable respuesta (numérica).}

\item{col_factor}{Vector de caracteres. Nombres de las columnas que actuarán como predictores (factores).}

\item{interaction}{Vector de caracteres (Opcional). Interacciones explícitas (ej. "FactorA:FactorB").}

\item{family}{String. Familia del GLM. Opciones: "Gamma", "Gaussian", "Quasipoisson".}

\item{dec}{Entero. Número de decimales para el redondeo. Por defecto 2.}

\item{output_name.}{Ruta/Nombre base para guardar los archivos de salida.}
}
\value{
Una lista con dos data frames: `summary_dem` (% Deviance Explicada) y `summary_raw` (Tabla completa con valores p).
}
\description{
Ajusta GLMs para múltiples parámetros simultáneamente, calcula el análisis de deviance (ANOVA) y genera tablas resumen con el porcentaje de deviance explicada (%DEM) y significancia.
}
