% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources_fn_abiotic.R
\name{get_norm_exceedance}
\alias{get_norm_exceedance}
\title{Calcular Excedencias de Normas Ambientales}
\usage{
get_norm_exceedance(
  data,
  norm_data,
  col_pars,
  col_value,
  col_group = NULL,
  output_name = "tabla_excedencias.tsv"
)
}
\arguments{
\item{data}{Data frame con los datos de mediciones.}

\item{norm_data}{Data frame con las normas. Debe contener columnas `limite_1` y `limite_2`.}

\item{col_pars}{String. Nombre de la columna de parámetros en `data` y `norm_data`.}

\item{col_value}{String. Nombre de la columna de valores en `data`.}

\item{col_group}{String (Opcional). Columna de agrupación (ej. "Zona", "Sitio").}

\item{output_name}{String. Nombre del archivo de salida (.tsv o .csv).}
}
\value{
Un data frame con el resumen de excedencias (invisible).
}
\description{
Evalúa los datos frente a una tabla de normas, calculando frecuencias y porcentajes de excedencia. Maneja lógica específica para rangos (ej. pH) y límites superiores.
}
