% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources_fn_abiotic.R
\name{get_summ_stats}
\alias{get_summ_stats}
\title{Resumen estadístico por parámetro}
\usage{
get_summ_stats(
  data,
  col_pars,
  col_value,
  col_bld,
  matrix,
  round = 2,
  output_name
)
}
\arguments{
\item{data}{Tabla de datos de entrada.}

\item{col_pars}{Nombre de la columna (como cadena) que identifica el parámetro.}

\item{col_value}{Nombre de la columna (como cadena) con los valores numéricos.}

\item{col_bld}{Nombre de la columna (como cadena) que indica registros bajo el límite de detección (BLD).}

\item{matrix}{Cadena que indica la matrix (por ejemplo, "agua", "sedimento") y controla el filtrado de parámetros granulométricos.}

\item{round}{Número de decimales para redondear los estadísticos (por defecto 2).}

\item{output_name}{Nombre de archivo de salida (CSV) donde se guardará la tabla de resultados (se genera también un TSV adicional).}
}
\description{
Calcula estadísticos descriptivos por parámetro
}
\examples{
\dontrun{
data <- readr::read_tsv("data_sedimento_inv_2025_clean.tsv")
col_pars <- "Sigla"
col_value <- "Valor"
matrix <- "sedimento"
round <- 4
fn_stats(data = data, col_pars = col_pars, col_value = col_value, data_pars = data_pars, matrix = matrix, round = round)
}
}
