% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources_fn_abiotic.R
\name{line_plot_by_time}
\alias{line_plot_by_time}
\title{Gráfico de Series Temporales con Detección de Outliers (DBSCAN)}
\usage{
line_plot_by_time(
  data,
  col_value,
  col_pars,
  col_x,
  col_site,
  col_year,
  matriz,
  col_units,
  col_group = NULL,
  ord_site = NULL,
  output_name,
  width = 8,
  height = 8,
  legend_label = "Sitio",
  pars_annot = NULL,
  n_col = 3,
  type_plot = "annotated",
  save_rds = FALSE
)
}
\arguments{
\item{data}{Data frame con los datos.}

\item{col_value}{String. Nombre de la columna con los valores numéricos.}

\item{col_pars}{String. Nombre de la columna del parámetro.}

\item{col_x}{String. Nombre de la columna eje X (ej. "Campaña" numérico o factor).}

\item{col_site}{String. Nombre de la columna de sitios.}

\item{col_year}{String. Nombre de la columna de año.}

\item{matriz}{String. Nombre de la matriz (para títulos).}

\item{col_units}{String. Nombre de la columna de unidades.}

\item{col_group}{String (Opcional). Columna para agrupación adicional (ej. Estrato).}

\item{ord_site}{Vector (Opcional). Orden de los sitios.}

\item{output_name}{String. Nombre del archivo de salida.}

\item{width}{Numérico. Ancho imagen.}

\item{height}{Numérico. Alto imagen.}

\item{legend_label}{String. Título de la leyenda de color.}

\item{pars_annot}{Vector (Opcional). Parámetros a excluir de la anotación de asteriscos BLD.}

\item{n_col}{Entero. Columnas en facet_wrap.}

\item{type_plot}{String. "annotated" (muestra outliers y BLD) o "clean" (filtra outliers y BLD).}

\item{save_rds}{Lógico. Guardar objeto ggplot.}

\item{xlabel}{String. Título del eje x.}

\item{ylabel}{String. Título del eje y.}
}
\description{
Genera gráficos de línea por sitio/campaña, detectando anomalías mediante DBSCAN y el método "Kneedle" para epsilon óptimo. Permite resaltar valores bajo límite de detección (BLD).
}
