% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources_fn_biotic.R
\name{plot_alluvial_taxa}
\alias{plot_alluvial_taxa}
\title{Plot Alluvial de Composición (Abundancia o Riqueza por Grupo)}
\usage{
plot_alluvial_taxa(
  data,
  col_site,
  ord_site = NULL,
  col_N,
  col_taxa_group,
  col_taxa = NULL,
  taxa_id,
  col_facet = NULL,
  ord_facet = NULL,
  metric = "abundance",
  legend_label = NULL,
  xlabel = "Sitios",
  top_n = 10,
  output_name,
  width = 10,
  height = 6,
  save_rds = FALSE
)
}
\arguments{
\item{data}{Data frame con los datos.}

\item{col_site}{String. Columna eje X.}

\item{ord_site}{Vector (Opcional). Orden de los sitios.}

\item{col_N}{String. Columna de abundancia.}

\item{col_taxa_group}{String. El grupo mayor que formará las cintas (ej. Clase, Familia).}

\item{col_taxa}{String (Opcional). La unidad menor a contar para riqueza (ej. Especie). Requerido si metric="richness".}

\item{taxa_id}{String. Nombre del grupo general para título.}

\item{col_facet}{String (Opcional). Columna para facetar.}

\item{ord_facet}{Vector (Opcional). Orden de los niveles del facet.}

\item{metric}{String. "abundance" (aplica Top N) o "richness" (Muestra TODOS los grupos, sin Top N).}

\item{legend_label}{String (Opcional). Título leyenda.}

\item{xlabel}{String. Título eje X.}

\item{top_n}{Entero. Filtra los N grupos principales (Solo aplica si metric="abundance").}

\item{output_name}{String. Nombre del archivo de salida (ej. "grafico.png").}

\item{width}{Numérico. Ancho imagen.}

\item{height}{Numérico. Alto imagen.}

\item{save_rds}{Lógico. Guardar RDS.}
}
\value{
Un objeto ggplot.
}
\description{
Crea un gráfico aluvial. Ordena la leyenda por magnitud y respeta la lógica de riqueza total sin filtrar si se solicita.
}
