% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources_fn_abiotic.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{Generar Boxplots con opción de pruebas estadísticas}
\usage{
plot_boxplot(
  data,
  col_x,
  col_y,
  col_pars,
  col_units,
  n_col = 3,
  output_name = "boxplot.png",
  width = 10,
  height = 8,
  levels_factor = NULL,
  title = "",
  label_x = "Sitios",
  label_y = "Valor",
  add_test = FALSE,
  save_rds = FALSE
)
}
\arguments{
\item{data}{Data frame con los datos.}

\item{col_x}{String. Nombre de la columna para el eje X (factor/grupos).}

\item{col_y}{String. Nombre de la columna para el eje Y (numérica).}

\item{col_pars}{String. Nombre de la columna de parámetros (para facet_wrap).}

\item{col_units}{String. Nombre de la columna de unidades.}

\item{n_col}{Entero. Número de columnas en el facet_wrap.}

\item{output_name}{String. Ruta/nombre del archivo de salida (png).}

\item{width}{Numérico. Ancho de la imagen.}

\item{height}{Numérico. Alto de la imagen.}

\item{levels_factor}{Vector. Niveles ordenados para el factor del eje X.}

\item{title}{String. Título del gráfico.}

\item{label_x}{String. Etiqueta eje X.}

\item{label_y}{String. Etiqueta eje Y.}

\item{add_test}{Lógico. Si es TRUE, agrega estadísticas (KW y/o Dunn). Por defecto FALSE.#' @return Guarda un archivo PNG y retorna el objeto ggplot invisiblmente.}

\item{save_rds}{Lógico. Si es TRUE, guarda también el objeto R (.rds). Por defecto FALSE.}
}
\description{
Crea boxplots faceteados. Opcionalmente realiza pruebas de Kruskal-Wallis y, si hay pocos grupos, pruebas post-hoc de Dunn.
}
