% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources_fn_abiotic.R
\name{plot_correlogram}
\alias{plot_correlogram}
\title{Correlograma de Spearman para Parámetros Ambientales}
\usage{
plot_correlogram(
  data,
  col_pars,
  col_site,
  col_value,
  matriz,
  col_group = NULL,
  col_factor = NULL,
  width = 6,
  height = 6,
  save_rds = FALSE
)
}
\arguments{
\item{data}{Un data.frame o tibble que contiene los datos.}

\item{col_pars}{String. Nombre de la columna que contiene los nombres de los parámetros (ej. "Parametro").}

\item{col_site}{String. Nombre de la columna que contiene los sitios/estaciones.}

\item{col_value}{String. Nombre de la columna que contiene los valores numéricos.}

\item{matriz}{String. Tipo de matriz (ej. "Agua", "Sedimento"). Si contiene "sedimento", se excluyen parámetros granulométricos.}

\item{col_group}{String (Opcional). Nombre de la columna para agrupación adicional. Por defecto NULL.}

\item{col_factor}{String (Opcional). Nombre de una columna de factor extra. Por defecto NULL.}

\item{width}{Numérico. Ancho de la imagen guardada en pulgadas. Por defecto 6.}

\item{height}{Numérico. Alto de la imagen guardada en pulgadas. Por defecto 6.}

\item{save_rds}{Lógico. Si es TRUE, guarda también el objeto R (.rds). Por defecto FALSE.}

\item{output_name}{String. Nombre archivo de salida.}
}
\value{
Un objeto ggplot2 con el correlograma. También guarda un archivo PNG en el directorio de trabajo.
}
\description{
Esta función procesa un data frame de parámetros ambientales, calcula correlaciones de Spearman y genera un gráfico de tipo "heatmap" (triángulo inferior)
}
\examples{
\dontrun{
plot_correlogram(
  data = mi_data,
  col_pars = "PARAMETRO",
  col_site = "ESTACION",
  col_value = "RESULTADO",
  matriz = "Agua de Mar"
)
}
}
