% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources_fn_abiotic.R
\name{plot_granulometria}
\alias{plot_granulometria}
\title{Graficar Distribución de Granulometría (Stacked Barplot)}
\usage{
plot_granulometria(
  data,
  col_pars,
  col_site,
  col_value,
  col_factor = NULL,
  col_rep = NULL,
  ord_site = NULL,
  levels_factor = NULL,
  width = 8,
  height = 6,
  output_name = "plot_granulometria.png",
  save_rds = FALSE
)
}
\arguments{
\item{data}{Data frame con los datos.}

\item{col_pars}{String. Nombre de la columna con los tipos de grano (ej. "Parametro").}

\item{col_site}{String. Nombre de la columna con los sitios.}

\item{col_value}{String. Nombre de la columna con el porcentaje/valor.}

\item{col_factor}{String (Opcional). Columna para agrupar en facetas (columnas).}

\item{col_rep}{String (Opcional). Columna para agrupar en facetas (filas).}

\item{ord_site}{Vector (Opcional). Orden específico para los sitios.}

\item{levels_factor}{Vector (Opcional). Orden específico para el factor de agrupación.}

\item{width}{Numérico. Ancho de la imagen guardada.}

\item{height}{Numérico. Alto de la imagen guardada.}

\item{output_name}{String. Nombre del archivo de salida.}

\item{save_rds}{Lógico. Si es TRUE, guarda también el objeto R (.rds). Por defecto FALSE.}
}
\value{
Un objeto ggplot.
}
\description{
Genera un gráfico de barras apiladas para visualizar la composición granulométrica por sitio. Permite agrupar por factores (ej. Zonas) y separar por réplicas.
}
