% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources_fn_biotic.R
\name{plot_treemap}
\alias{plot_treemap}
\title{Generar Treemap de Composición Taxonómica}
\usage{
plot_treemap(
  data,
  col_site = NULL,
  ord_site = NULL,
  col_N,
  col_facet = NULL,
  ord_facet = NULL,
  col_zone = NULL,
  ord_zone = NULL,
  taxa_id,
  taxa_group,
  width = 8,
  height = 10,
  output_name = "plot_treemap.png"
)
}
\arguments{
\item{data}{Data frame con los datos.}

\item{col_site}{String (Opcional). Nombre de la columna de sitios. (Anteriormente `col_sitio`).}

\item{ord_site}{Vector (Opcional). Orden de los niveles de sitios.}

\item{col_N}{String. Nombre de la columna de abundancia o conteo. (Anteriormente `col_N`).}

\item{col_facet}{String (Opcional). Nombre de la columna para facetar.}

\item{ord_facet}{Vector (Opcional). Orden de los niveles del facet.}

\item{col_zone}{String (Opcional). Nombre de la columna de zonas. (Anteriormente `col_zona`).}

\item{ord_zone}{Vector (Opcional). Orden de los niveles de zonas.}

\item{taxa_id}{String. Texto identificador del taxón superior (ej. "Familia X") para el título.}

\item{taxa_group}{String. Nombre de la columna con el grupo taxonómico a graficar.}

\item{width}{Numérico. Ancho de la imagen. Default 8.}

\item{height}{Numérico. Alto de la imagen. Default 10.}

\item{output_name}{String. Nombre del archivo de salida.}
}
\value{
Un objeto ggplot (invisible).
}
\description{
Crea un gráfico Treemap basado en abundancias. Soporta facetado, agrupación por sitio/zona y formateo específico de etiquetas.
}
